WIP поменял тип координат с double на int; добавил классы для работы с координатами

+++ src/main/java/ru/mainmayhem/arenaofglory/data/entities/Coordinates.kt
@@ -1,7 +1,10 @@
 package ru.mainmayhem.arenaofglory.data.entities
 
 data class Coordinates(
-    val x: Double,
-    val y: Double,
-    val z: Double
+    //возрастает с запада на восток
+    val x: Int,
+    //возрастает снизу вверх
+    val y: Int,
+    //возрастает с севера на юг
+    val z: Int
 )
