+++ src/main/java/ru/mainmayhem/arenaofglory/data/local/database/tables/exposed/WaitingRoomCoordinates.kt
@@ -4,9 +4,9 @@ import org.jetbrains.exposed.sql.Table
 
 object WaitingRoomCoordinates: Table() {
     val type = varchar("id", 25)
-    val x = double("x")
-    val y = double("y")
-    val z = double("z")
+    val x = integer("x")
+    val y = integer("y")
+    val z = integer("z")
     override val tableName: String = "waiting_room_coordinates"
     override val primaryKey: PrimaryKey = PrimaryKey(type)
 }
